+++ CharlieBackend.Api/Controllers/LessonsController.cs
@@ -102,21 +102,7 @@ namespace CharlieBackend.Api.Controllers
             return lessons.ToActionResult();
         }
 
-        /// <summary>
-        /// Returns list of lessons  for mentor
-        /// </summary>
-        /// <param name="id"></param>
-        /// <response code="200">Successful return of lessons list of given student</response>
-        [SwaggerResponse(200, type: typeof(IList<LessonDto>))]
-        [Authorize(Roles = "Admin, Mentor, Secretary")]
-        [HttpGet("mentor/{id}")]
-        public async Task<ActionResult<List<LessonDto>>> GetAllLessonsForMentor(long id)
-        {
-            var lessons = await _lessonService.GetAllLessonsForMentor(id);
-
-            return lessons.ToActionResult();
-        }
-
+        
         /// <summary>
         /// Updates given lesson
         /// </summary>
