+++ CharlieBackend.Business/Services/LessonService.cs
@@ -74,15 +74,6 @@ namespace CharlieBackend.Business.Services
 
         public async Task<Result<IList<LessonDto>>> GetAllLessonsForMentor(long mentorId)
         {
-            if (mentorId == default)
-            {
-                return Result<IList<LessonDto>>.GetError(ErrorCode.ValidationError, "Write Mentod Id please");
-            }
-            if (await _unitOfWork.MentorRepository.GetByIdAsync(mentorId) == null)
-            {
-                return Result<IList<LessonDto>>.GetError(ErrorCode.ValidationError, $"Mentod with id {mentorId} is not Found");
-            }
-
             var lessons = await _unitOfWork.LessonRepository.GetAllLessonsForMentor(mentorId);
 
             return Result<IList<LessonDto>>.GetSuccess(_mapper.Map<IList<LessonDto>>(lessons));
